refactor Attribute folder
